% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ch_list_add.R
\name{ch_list_add}
\alias{ch_list_add}
\title{Add an item to a list}
\usage{
ch_list_add(id, item = "item", plural = NULL)
}
\arguments{
\item{id}{character, tag to prepend to the input and output id's}

\item{item}{character, for the ui, what to call the item being added - default is "item"}

\item{plural}{character, for the ui, what to call the plural of the item being added
                 - default is \code{paste0(item, "s")}}
}
\value{
list containing \code{ui_controller}, \code{ui_view}, and \code{server_model}
}
\description{
Creates a list of shiny objects to encapsulate the operation of adding an item to a list.
}
\details{
The list of shiny objects will contain:

\describe{
  \item{\code{ui_controller}}{\code{shiny::taglist} of ui elements for the controller}
  \item{\code{ui_view}}{\code{shing::taglist} of ui elements for the view}
  \item{\code{server_model}}{function with reactive code}
}

The list returned by this function has to be available to both the ui and the server.
If not using the \code{shinyApp} formulation, perhaps \code{global.R} could be useful.

The list \code{ui_controller} will have members:

\describe{
 \item{\code{name}}{\code{shiny::textInput} used to specify the name of the item to be added to the list}
 \item{\code{add}}{\code{shiny::actionButton} used to invoke the action of adding the item to the list}
}

The list \code{ui_view} will have members:

\describe{
 \item{\code{status}}{\code{shiny::verbatimTextOutput} showing if it is possible to add to the list}
}

The function \code{server_model()} will be called from your server function.
Its arguments are:

\describe{
 \item{\code{input, output, session}}{input, output, session values passed from your server function}
 \item{\code{rctval_source, item_source}}{
   \code{shiny::reactiveValues} object, character string

\code{rctval_source[[item_source]]} an object of some sort to be added to the list.
 }
 \item{\code{rctval_list, item_list}}{
   \code{shiny::reactiveValues} object, character string

The default value for \code{rctval_list} is \code{rctval_source}.
   \code{rctval_list[[item_list]]} is expected to be a list, to which the object will be added
 }
}
}

