% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ch_list_select.R
\name{ch_list_select}
\alias{ch_list_select}
\title{Select an item from a list}
\usage{
ch_list_select(id, item = "item", plural = NULL)
}
\arguments{
\item{id}{character, tag to prepend to the input and output id's}

\item{item}{character, for the ui, what to call the item being added - default is "item"}

\item{plural}{character, for the ui, what to call the plural of the item being added
                 - default is \code{paste0(item, "s")}}
}
\value{
list containing \code{ui_controller}, \code{ui_view}, and \code{server_model}
}
\description{
Creates a list of shiny objects to encapsulate the operation of selecting an item from a list.
}
\details{
The list of shiny objects will contain:

\describe{
  \item{\code{ui_controller}}{\code{shiny::taglist} of ui elements for the controller}
  \item{\code{ui_view}}{\code{shing::taglist} of ui elements for the view}
  \item{\code{server_model}}{function with reactive code}
}

The list returned by this function has to be available to both the ui and the server.
If not using the \code{shinyApp} formulation, perhaps \code{global.R} could be useful.

The list \code{ui_controller} will have members:

\describe{
 \item{\code{item}}{\code{shiny::selectInput} used to select the name of the item to be copied from the list}
}

The list \code{ui_view} will have members:

\describe{
 \item{\code{status}}{\code{shiny::verbatimTextOutput} showing what has been selected from the list}
}

The function \code{server_model()} will be called from your server function.
Its arguments are:

\describe{
 \item{\code{input, output, session}}{input, output, session values passed from your server function}
 \item{\code{rctval_list, item_list}}{
   \code{shiny::reactiveValues} object, character string

\code{rctval_list[[item_list]]} is expected to be a list,
   from which the selected item will be copied
 }
 \item{\code{rctval_dest, item_dest}}{
   \code{shiny::reactiveValues} object, character string

The default value for \code{rctval_dest} is \code{rctval_list}.
   \code{rctval_dest[[item_dest]]} is where the selected item will be copied.
 }
}
}

