% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ch_read_delim.R
\name{ch_read_delim}
\alias{ch_read_delim}
\title{Creates a collection of shiny objects to wrap the \code{readr::read_delim} function.}
\usage{
ch_read_delim(id)
}
\arguments{
\item{id}{character, tag to prepend to the input and output id's}
}
\value{
list containing \code{ui_controller}, \code{ui_view}, and \code{server_model}
}
\description{
The list will contain:
}
\details{
\describe{
  \item{\code{ui_controller}}{\code{shiny::taglist} of ui elements for the controller}
  \item{\code{ui_view}}{\code{shing::taglist} of ui elements for the view}
  \item{\code{server_model}}{function with reactive code}
}

The list returned by this function has to be available to both the ui and the server.
If not using the \code{shinyApp} formulation, perhaps \code{global.R} could be useful.

The list \code{ui_controller} will have members:

\describe{
 \item{\code{file}}{\code{shiny::fileInput} used to choose the file to upload}
 \item{\code{delim}}{\code{shiny::selectInput} used to select the delimiter character }
 \item{\code{decimal_mark}}{\code{shiny::selectInput used to select the decimal-mark character}}
 \item{\code{tz_file}}{\code{shiny::selectInput used to select the timezone used in the file}}
 \item{\code{tz_file_model}}{\code{shinyBS::bsModal used for timezone help}}
 \item{\code{tz_location}}{\code{shiny::selectInput used to select the timezone at the location}}
 \item{\code{tz_location_model}}{\code{shinyBS::bsModal used for timezone help}}
}

Note that the `tz` input will serve as an argument to `readr::locale`;
its meaning depends on the context of what it's parsing. See the readr
documentation for more details.

The list \code{ui_view} will have members:

\describe{
 \item{\code{text}}{\code{shiny::htmlOutput} showing a preview of the first few lines of the text file}
 \item{\code{data}}{\code{shiny::htmlOutput} showing a glimpse of the parsed dataframe}
}

The function \code{server_model()} will be called from your server function.
Its arguments are:

\describe{
 \item{\code{input, output, session}}{input, output, session values passed from your server function}
 \item{\code{rctval_data, item_data}}{
   \code{shiny::reactiveValues} object, character string.
   The parsed dataframe will be placed in \code{rctval_data[[item_data]]}.
 }
}
}

