% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{cl_file_copy}
\alias{cl_file_copy}
\title{Closure for file-downloading}
\usage{
cl_file_copy(file_source)
}
\arguments{
\item{file_source}{character describing path to file to be copied}
}
\value{
function
}
\description{
A closure to provide a function to download a file,
by copying it from the specified place. Useful for
\code{shiny::downloadHandler()}.
}

