% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{df_set_tz}
\alias{df_set_tz}
\title{df_set_tz}
\usage{
df_set_tz(df, tz = "UTC")
}
\arguments{
\item{df}{dataframe}

\item{tz}{Olson timezone}
}
\value{
dataframe
}
\description{
Change the timezone of all the POSIX.ct object to tz
}

