% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ch_list_select.R
\name{rctval_select}
\alias{rctval_select}
\title{Creates a list of shiny objects to encapsulate the operation of
selecting a single item from a list.}
\usage{
rctval_select(id, item = "item", plural = NULL)
}
\arguments{
\item{id}{character, tag to prepend to the input and output id's}

\item{item}{character, for the ui, what to call the item being added - default is "item"}

\item{plural}{character, for the ui, what to call the plural of the item being added
                 - default is \code{paste0(item, "s")}}
}
\value{
list containing \code{ui_controller}, \code{ui_view}, and \code{server_model}
}
\description{
The list of shiny objects will contain:
}
\details{
\describe{
  \item{\code{ui_controller}}{\code{shiny::taglist} of ui elements for the controller}
  \item{\code{ui_view}}{\code{shing::taglist} of ui elements for the view}
  \item{\code{server_model}}{function with reactive code}
}

The list returned by this function has to be available to both the ui and the server.
If not using the \code{shinyApp} formulation, perhaps \code{global.R} could be useful.

The list \code{ui_controller} will have members:

\describe{
 \item{\code{time}}{\code{shiny::selectInput} used to choose the variable for the x axis}
 \item{\code{y1}}{\code{shiny::selectInput} used to choose the variable for the y1 axis}
 \item{\code{y2}}{\code{shiny::selectInput} used to choose the variable for the y2 axis}
}

The list \code{ui_view} will have members:

\describe{
 \item{\code{dygraph}}{\code{dygraphs::dygraphOutput} showing a preview of the first few lines of the text file}
}

The function \code{server_model()} will be called from your server function.
Its arguments are:

\describe{
 \item{\code{input, output, session}}{input, output, session values passed from your server function}
 \item{\code{rctval_data, item_data}}{
   \code{shiny::reactiveValues} object, character string.
   \code{rctval_data[[item_data]]} is expected to be a dataframe, input for dygraph.
 }
 \item{\code{rctval_dyopt, item_dyopt}}{
   \code{shiny::reactiveValues} object, character string.
   The default value for \code{rctval_dyopt} is \code{rctval_dyopt}.
   \code{rctval_data[[item_data]]} is expected to be a list of dygraph options,
   see documentation for \code{dygraphs::dyOptions}
 }
}
}

