% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tbl_df_read_delim.R
\name{tbl_df_read_delim}
\alias{tbl_df_read_delim}
\title{tbl_df_read_delim}
\usage{
tbl_df_read_delim(id)
}
\arguments{
\item{id}{character, tag to prepend to the input and output id's}
}
\value{
list containing \code{ui_control}, \code{ui_response}, and \code{srv_function}
}
\description{
Creates a collection of shiny objects to manage the parsing of a delimited file.
}
\details{
The list will contain:

\itemize{
  \item \code{ui_control} \code{shiny::taglist} of control items
  \item \code{ui_response} \code{shing::taglist} of response items
  \item \code{srv_function} a function with reactive code - this function takes a \code{reactiveValues} as its arguement
}

The list returned by the factory has to be available to both the ui and the server. If not using the \code{shinyApp}
formulation, perhaps \code{global.R} could be useful.

The component \code{srv_function} will be called in the server function, using the
particular \code{reactiveValue} you wish to associate with the "thing".
}

