% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tbldf_read_delim.R
\name{tbldf_read}
\alias{tbldf_read}
\title{tbldf_read}
\usage{
tbldf_read(file, delim = ",", tz_file = "UTC", tz_location = "UTC")
}
\arguments{
\item{file}{Passed to \code{readr::read_delim}. Either a path to a file, a connection, or literal data (either a single string or a raw vector).}

\item{delim}{Passed to \code{readr::read_delim}. Delimiter.}

\item{tz_file}{Timezone used to express the time in the file, see \code{lubridate::olson_time_zones()}}

\item{tz_location}{Timezone used to express the time at the location, see \code{lubridate::olson_time_zones()}}
}
\value{
\code{dplyr::tbl_df()} Dataframe-like structure
}
\description{
Given some text, a delimiter, and tz attributes, return a data-frame
}

