% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wx_ames.R
\docType{data}
\name{wx_ames}
\alias{wx_ames}
\title{Weather observations at Ames - Jan. 2014}
\format{A data frame with 983 rows and 19 variables:
\describe{
  \item{date}{}
  \item{temp}{}
  \item{dew_pt}{}
  \item{hum}{}
  \item{wind_spd}{}
  \item{wind_gust}{}
  \item{dir}{}
  \item{vis}{}
  \item{pressure}{}
  \item{wind_chill}{}
  \item{heat_index}{}
  \item{precip}{}
  \item{cond}{}
  \item{fog}{}
  \item{rain}{}
  \item{snow}{}
  \item{hail}{}
  \item{thunder}{}
  \item{tornado}{}
}}
\source{
\url{http://www.wunderground.com/}
}
\usage{
wx_ames
}
\description{
A dataset containing hourly weather observations at Ames, Iowa (US) for
the month of January 2014. Data obtained from Weather Undergound's API.
}
\keyword{datasets}

